382: Swap around the getLoggingEventBuilder to the local logger
